Code refactorign to impreave readability
